+++ service/internal/auth/authn.go
@@ -114,6 +114,8 @@ func NewAuthenticator(ctx context.Context, cfg Config, logger *logger.Logger, we
 	if err != nil {
 		return nil, err
 	}
+	// Assign configured public_client_id
+	oidcConfig.PublicClientID = cfg.PublicClientID
 
 	// If the issuer is different from the one in the configuration, update the configuration
 	// This could happen if we are hitting an internal endpoint. Example we might point to https://keycloak.opentdf.svc/realms/opentdf